Code/backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
ibm-watson-machine-learning>=1.0.0
python-dotenv>=1.0.0
